Make set_preview_files_for_entities reusable
