+++ zou/app/services/preview_files_service.py
@@ -11,6 +11,7 @@ from zou.app.models.preview_file import PreviewFile
 from zou.app.models.project import Project
 from zou.app.models.project_status import ProjectStatus
 from zou.app.models.task import Task
+from zou.app.models.task_type import TaskType
 from zou.app.services import names_service, files_service
 from zou.utils import movie
 from zou.app.utils import (
@@ -418,3 +419,115 @@ def get_running_preview_files():
         )
         results.append(result)
     return results
+
+
+def set_preview_files_for_entities(entities):
+    """
+    Retrieve all preview files related to given entities.
+    Add to each entity a dict with task as keys and preview list as values.
+    """
+    entity_ids = []
+    for entity in entities:
+        if "id" not in entity:
+            entity_id = entity.get("shot_id", entity.get("entity_id", None))
+            if entity_id is not None:
+                entity_ids.append(entity_id)
+                entity["id"] = entity_id
+        else:
+            entity_ids.append(entity["id"])
+    previews = {}
+    preview_file_map = {}
+
+    preview_files = (
+        PreviewFile.query.join(Task)
+        .join(TaskType)
+        .filter(Task.entity_id.in_(entity_ids))
+        .order_by(TaskType.priority.desc())
+        .order_by(TaskType.name)
+        .order_by(PreviewFile.revision.desc())
+        .order_by(PreviewFile.created_at)
+        .add_column(Task.task_type_id)
+        .add_column(Task.entity_id)
+        .all()
+    )
+
+    is_pictures = False
+    for (preview_file, task_type_id, entity_id) in preview_files:
+        entity_id = str(entity_id)
+        task_type_id = str(task_type_id)
+        if entity_id not in previews:
+            previews[entity_id] = {}
+
+        if task_type_id not in previews[entity_id]:
+            previews[entity_id][task_type_id] = []
+
+        if preview_file.extension == "png":
+            is_pictures = True
+
+        task_id = str(preview_file.task_id)
+        preview_file_id = str(preview_file.id)
+
+        light_preview_file = {
+            "id": preview_file_id,
+            "revision": preview_file.revision,
+            "extension": preview_file.extension,
+            "status": str(preview_file.status).lower(),
+            "annotations": preview_file.annotations,
+            "created_at": fields.serialize_value(preview_file.created_at),
+            "task_id": task_id,
+        }  # Do not add too much field to avoid building too big responses
+        previews[entity_id][task_type_id].append(light_preview_file)
+        preview_file_map[preview_file_id] = light_preview_file
+
+    if is_pictures:
+        for entity_id in previews.keys():
+            for task_type_id in previews[entity_id].keys():
+                previews[entity_id][task_type_id] = mix_preview_file_revisions(
+                    previews[entity_id][task_type_id]
+                )
+
+    for entity in entities:
+        if str(entity["id"]) in previews:
+            entity["preview_files"] = previews[str(entity["id"])]
+        else:
+            entity["preview_files"] = []
+
+        try:
+            preview_file_id = entity["preview_file_id"]
+            preview_file = preview_file_map.get(preview_file_id, None)
+            if preview_file is not None:
+                entity.update(
+                    {
+                        "preview_file_id": preview_file["id"],
+                        "preview_file_extension": preview_file["extension"],
+                        "preview_file_status": preview_file["status"],
+                        "preview_file_annotations": preview_file[
+                            "annotations"
+                        ],
+                        "preview_file_task_id": preview_file["task_id"],
+                        "preview_file_previews": preview_file["previews"],
+                    }
+                )
+        except Exception as e:
+            print(e)
+
+    return preview_file_map
+
+
+def mix_preview_file_revisions(preview_files):
+    """
+    The goal here is to group preview files with same revision in a single
+    preview file, which encapsulates other preview_files.
+    """
+    revision_map = {}
+    result = []
+    for preview_file in preview_files:
+        revision = preview_file["revision"]
+        if revision not in revision_map:
+            preview_file["previews"] = []
+            revision_map[revision] = preview_file
+            result.append(preview_file)
+        else:
+            parent_preview_file = revision_map[revision]
+            parent_preview_file["previews"].append(preview_file)
+    return result
